+++ src/main/java/me/pajic/accessorify/util/compat/CompatFlags.java
@@ -11,6 +11,7 @@ public class CompatFlags {
     public static final boolean FABRIC_SEASONS_LOADED = FabricLoader.getInstance().isModLoaded("seasons");
     public static final boolean FABRIC_SEASONS_EXTRAS_LOADED = FabricLoader.getInstance().isModLoaded("seasonsextras");
     public static final boolean ARS_ELIXIRUM_LOADED = FabricLoader.getInstance().isModLoaded("elixirum");
+    public static final boolean NETHERITE_EXTRAS_LOADED = FabricLoader.getInstance().isModLoaded("netheriteextras");
     public static final boolean ADDITIONAL_LANTERNS_LOADED = FabricLoader.getInstance().isModLoaded("additionallanterns");
     public static final boolean REINFORCED_SHULKERS_LOADED = FabricLoader.getInstance().isModLoaded("reinfshulker");
     public static final boolean IMMERSIVE_OVERLAYS_LOADED = FabricLoader.getInstance().isModLoaded("immersiveoverlays");